Refactor: seperate input components from steps
