--- src/components/Auth/Signup/Step/Step.module.scss
@@ -1,189 +1,0 @@
-@use "@/sass" as *;
-
-.container {
-  position: relative;
-
-  label {
-    display: block;
-
-    @include typography(tabLabel);
-    color: $neutral900;
-  }
-
-  small {
-    width: 100%;
-    position: absolute;
-    bottom: -4px;
-    left: 0;
-
-    @include typography(captionSmall);
-    color: $danger300;
-  }
-
-  .email,
-  .sertCode,
-  .emailResert,
-  .password,
-  .passwordConfirm,
-  .nickname {
-    position: relative;
-    margin-top: 20px;
-
-    &:focus-within {
-      input {
-        outline: none;
-        border-width: 1px;
-        border-color: $primary300;
-
-        &.error {
-          border-color: $danger300;
-        }
-      }
-
-      .removeBtn {
-        display: block;
-      }
-    }
-  }
-
-  .input {
-    width: 100%;
-    margin-top: 6px;
-    margin-bottom: 24px;
-    padding: 16px;
-    border: 1px solid $neutral200;
-    border-radius: 8px;
-
-    @include typography(bodyLarge);
-    color: $neutral900;
-
-    &::placeholder {
-      color: $neutral400;
-    }
-
-    &.error {
-      border-color: $danger300;
-    }
-
-    &:disabled {
-      color: $neutral300;
-    }
-  }
-
-  .removeBtn {
-    display: none;
-    position: absolute;
-    top: 47px;
-    right: 14px;
-    cursor: default;
-  }
-
-  .userEmail {
-    &__label {
-      @include typography(tabLabel);
-      color: $neutral900;
-    }
-
-    &__email {
-      @include typography(titleSmall);
-    }
-  }
-
-  .sertCode {
-    position: relative;
-
-    input {
-      margin-bottom: 12px;
-    }
-
-    &__label {
-      margin-top: 32px;
-    }
-
-    &__due {
-      position: absolute;
-      top: 45px;
-      right: 14px;
-
-      @include typography(bodyLarge);
-      color: $primary300;
-
-      &.timeout {
-        color: $neutral300;
-      }
-    }
-  }
-
-  .emailResert {
-    display: flex;
-    gap: 8px;
-    margin-top: 0;
-
-    @include typography(captionSmall);
-
-    &__timein {
-      color: $neutral400;
-    }
-    &__timeout {
-      color: $danger300;
-    }
-    &__btn {
-      @include typography(captionSmall);
-
-      &:hover {
-        font-weight: 700;
-      }
-    }
-  }
-
-  .image {
-    position: relative;
-    padding-top: 16px;
-    height: 8.8rem;
-    margin-bottom: 32px;
-
-    &__label {
-      position: absolute;
-      left: 50%;
-      transform: translateX(-50%);
-
-      background-position: center;
-      background-repeat: no-repeat;
-      background-size: cover;
-
-      width: 7.2rem;
-      height: 7.2rem;
-      padding: 8px;
-      border-radius: 50%;
-
-      &__camera {
-        background-color: $neutral100;
-        border: 1px solid $neutral200;
-        border-radius: 50%;
-
-        width: 2.6rem;
-        height: 2.6rem;
-        padding: 4px;
-
-        position: absolute;
-        bottom: 0;
-        right: 0;
-      }
-    }
-
-    &__input {
-      width: 100%;
-      height: 100%;
-      visibility: hidden;
-    }
-  }
-
-  .nickname {
-    &__disc {
-      @include typography(captionSmall);
-      color: $neutral400;
-      margin-top: 2px;
-      margin-bottom: 12px;
-    }
-  }
-}
